Fixes #7918.  Error with single splat method with keyword arguments
Problem was that we were seeing an empty kwarg marked from
ruby2_keyword_hash and omitting it because it was a kwarg.  The
logic doing this was for a failing test in MRI test suite that
was passing that empty kwarg using a splat.  The fix was to
only pass this along if it is a non-splat caller.