Fix #7815. NullPointerException for compiled code with for loop
AOT code encodes IR into the .class file and it was not correctly
adjusting local variable operands to account for `for` loops.  The
code before the fix would count all possible `for` loops and not
just the `for` loops between the current accessing scope and the
destination scope.

Added a single test which has a complicated set of variable accesses
which should exercise the depth adjustment which was fixed in this
commit.