Prepend internal marker to internal sources
This allows tools to omit any of our internal Ruby-based kernel
sources when processing backtraces and source_locations. The
filtering here is only done in source_location and backtrace
generation currently, with some extra logic during backtrace
mining to skip internal frames for e.g. Kernel#warn.

I only did this filtering at the highest levels, when the filename
is actually about to be used and returned to a Ruby consumer, to
avoid any internal usage of these filenames breaking due to the
new prefix.

Fixes #6430