Move more SwitchPoints before guarded values
The SwitchPoint used to invalidate a given value should always be
acquired prior to retrieving the value. This ensures that value
will always be guarded by its corresponding SwitchPoint, or if the
value changes between the two accesses, the value will be
associated with a prior invalid SwitchPoint and very shortly end
up back in the fallback to get the new value.

With several of these changes, the SwitchPoint was acquired after
the guarded value, meaning we might accidentally associate an old
value with a new SwitchPoint. If that new SwitchPoint is never
invalidated, we will cache the old value forever.

This relates to global variable fixes in prior commits as part of
https://github.com/jruby/jruby/pull/8003.