Don't use "ready to throw" utility methods
The Ruby.newBlahError methods are expected to return a fully-
prepared RaiseException with stack trace and current $! set. If
these methods are used to create a RubyException for throwing on
another thread, the current thread's stack trace and $! will be
used incorrectly.

Instead, I start the process of adding direct constructors for
Ruby exceptions, starting with RubyIOError.newIOError, which will
return just the Ruby exception object without any preparation for
throwing.