Narrow the glob to try to make this more reliable
Even with a 20 second GC loop this still occasionally fails on
Windows. This means one of two things:

* 20 seconds of GCs with 0.1s pause between each is insufficient
  to ensure that the tempfiles have been cleaned up. If so, there
  may be no amount of time we can run to ensure this test passes.
* Windows may be depositing or detecting other files in this temp
  dir and failing as a result.

The former can't be fixed; if that is the case, and this test
works correctly on other platforms, we may have to accept that it
can't be run reliably on Windows (and exclude it there).

I attempt to improve the latter here by only searching for files
that will match the requested temp file name.