Re-set handler before executing
Doing this re-set at the bottom led to a situation where the
signal handler thread – running in parallel – might overlap with
another signal trap, and while finishing up this handler logic it
could wipe out that trap's handler. This led to unpredictable
hangs in some signal-related tests and specs as detailed in #8007.