Separate two types of internal check
* Starting with uri:classloader:/jruby/kernel are filenames that
  come from JRuby's internal Ruby sources. They should gain the
  internal prefix.
* Starting with internal prefix indicates an internal source that
  may be appropriate to exclude under certain conditions.

We check both of these at trace-construction time because the
relevant rubyspec shows the behavior to also filter out user-
provided filenames with the same internal prefix, which is used
to test the masking and removal of these lines. In order to
handle both true internal filenames and user-provided internal-
prefixed filenames, both must be checked for omission here.

I don't agree with the side effect of being able to mask any trace
line, but I think that train has sailed.