Fixes language spec involving block calling #to_ary
The issue is if we try to call to_ary during block dispatch and
respond_to? does not exist on the object it should return the
original object.  We were failing because we rightly called
method_missing but because m_m did not handle to_ary we just
raised NoMethodError and not the original object.