Fix long-standing issue with block dispatch and spreading.
In the presence of only opt-args a single array value will spread
but we have not.  I adjusted our isSpreadable method to "form-fit"
to make everything pass.  An interesting note that part of my
fit was to realize kwrest was special and not in spreadable but
did happen to see it in the same code JIT had.  both interp
and JIT both call into the same method now.