Try to find class resource to avoid exception
The exception thrown here when we cannot find a class can be very
expensive at boot time. For most third-party extensions, this
exception will be raised twice for every bound method, due to
double-checking the classloader under lock a second time. By first
looking for a .class resource, we can avoid the expensive
exception if it is unlikely to succeed.