Fix String#unpack from using capacity() and causing wrong error
JIT and new Prism parser were right-sizing Strings whereas the
current parser will create larger than actual string length mostly.
This lead to a problem where checkLimit in original unpack code
would examine capacity and pass the test even though capacity
is not valid data.

The solution is to raise the proper error and use actual encoding
limit.  the original checkLimit error message is for STRTOUL in MRI
in processing the number after the '@' and that was not what we were
doing.  I am unable to get that RangeError to occur in MRI so I am
not sure when it is possible?