This message changed at some point
I can see a change in 2001, in ruby/ruby@dc98adf839, but I know
we did most of JRuby implementation after that. I suspect the
message was inconsistent for some years and eventually came back
to this version throughout CRuby.

In any case, the old message no longer exists anywhere in the
CRuby code base.