Update note about readline
readline 0.0.4 does not depend upon the readline-ext gem, and will
load reline if it is not installed. This is a final step toward
using reline exclusively and having no dependency on libreadline.

The equivalent in JRuby would be to install a similar stub,
ideally from the same CRuby gem, which attempts to load
jruby-readline. If that gem is not installed fall back on reline.
The implications of removing jruby-readline are unclear, since it
is still a dependency of the jirb_swing application.