Update strscan to fix improper sharing
The issue arises when the `StringScanner` string is being modified
while being scanned. In such a case, that string's buffers will be
improperly shared, forcing a new buffer to be created each time it
is modified. Meanwhile the ever-growing buffers are held in memory
by the improperly shared return values.

This was fixed by https://github.com/ruby/strscan/issues/84 and
strscan is updated in this PR.