Rework visibility checks for Java integration
In #8061 we found that Gradle is opening several core JDK packages
to support the needs of Groovy. Unfortunately this causes some
normally-inaccessible methods to be accessible to JRuby's Java
integration layer. In the case of HashSet, a package-private
`map` method gets bound that conflicts with the Ruby version,
leading to the bug reported in #8061.

Gradle's behavior is unsanitary, but ultimately the fix we have
settled on here is to never try to set package-private methods
accessible since they are truly not intended for consumption
outside the package. That fixes this issue without addressing the
larger problem of Gradle opening up core JDK packages.

Fixes #8061