Implement blocked waiter interruption with timeout
The timeout version of this logic was copied as-is from the Doug
Lea implementation, but did not have interruption logic to handle
closing a SizedQueue while there are threads waiting to push. This
patch incorporates the same changes made by @thedarkone for Queue
in 7f8c26fab7288a1dfb736bae3e931658d768f843.