Redo java commandline calculation to work with newer Java's
The old mechanism (which is still in this) only works for Java 8 and earlier.
It walks all ClassLoaders and will cast to a unpublished type to ask for
the URL for the jar the classloader represents.  Java 9+ cannot do this.

This fix only addresses the main Java jar by asking for a Resource
for our Main class then getting a URI which we can make into a valid
path to our jar.

What this does not do is walk all CLs.  It instead assumes all cp passed
into the original Java process will work for a nested one.  This may not
work for 9+ but it is already broken if so.  Java 8 continues to work the
same way it always has.