Remove deprecated 18/19/20 methods
This removes methods with the "18", "19", or "20" suffix
indicating that they are associated with those versions of Ruby.
Where there existed such methods that were not previously
deprecated, new deprecation was added but the methods were not
removed. Any remaining callers of these methods were modified to
call the non-suffixed version.