Add a benchmark for block_given? vs defined?(yield)
The latter form avoids any caller frame requirement since it just
takes the block in hand and checks if it is given. The former
form has typically required a caller frame, since it performs a
normal fcall to block_given? which then needs to be able to see
the caller's received block.