Add optimized logic for __method__ and __callee__
The __method__ and __callee__ methods currently force a frame due
to the default implementation needing to access the caller's frame
to get the method name or compound method name passed on the
stack. This optimization moves these calls to a specialized
instruction that can access the method/callee name directly
without needing a frame, improving performance of methods that use
this behavior.