Simplify fixnum and string ranges
Fixnum and String ranges can be reduced in complexity by doing
all of the construction in one go rather than each piece
individually. A Range of fixnum..fixnum now uses indy to pass the
long values to the bootstrap, avoiding bytecode to construct the
fixnum objects. A Range of string..string embeds the bytelist and
CR into the bootstrap parameters, for the same result. Both also
have simplified forms in the non-indy JIT mode.

From ideas list in #7588.