Specialize dstring paths using indy
This patch introduces a new indy call site for dynamic strings.
Instead of using frozen strings for the constant bits and doing
all construction and concatenation in the call-side bytecode, this
version embeds all constant strings into the indy call site and
passes the dynamic arguments only. This moves all concatenation
logic into the call site, where it can be specialized for small
arities or dumped into a generic large-arity string builder.

Note about consecutive static strings:

These probably should not happen, but due to the way we propagate
literal frozen strings it's possible to have multiple static
string pieces in a row. We will want to look at the optimizations
done in LocalOptimizationPass to re-calcuate the condensed set of
operands (combining consecutive string elements) to ensure we are
not wasting cycles on concatenating multiple static strings that
could be a single operation.