Reverse on previous decision and add castToXXX
I realized having throw be in the error methods had too many tradeoffs.
Beyond Java not being able to realize those methods unconditionally
throws (making some return null;  // not reached lines).  It also lacks
the ablity to compose:

```java
withException(context, typeError(context, a, "Array"), e);
```

Doing with this typeError raising is not possible and you would have
to either not have composable methods or make a combinatorial set of
methods which do the same thing.  This commit also has some oddities
for the new castTo but I think there are ok.

The first one is castTo is effectively replacing:

```java
if (a instanceof RubyArray) throw TypeError(context, a, "Array");
RubyArray b = (RubyArray) a;
```

with:

```java
RubyArray b = castToArray(context, a);
```

Looks nice but:
```java
RubyArray b = castToArray(getRuntime.getCurrentContext(), a);
```

Notice we pay price of looking up context when we may not do that in
the original snippet?  I think worst-case you can still do the original
snippet but in normal cases we continue to work on propagating ThreadContext
everywhere and this again will be no cost.

The second oddity is:

```java
if (a instanceof RubyArray) throw TypeError(context, a, str(runtime, ...));
RubyArray b = (RubyArray) a;
```

This form is accepted for castTo as well:

```java
RubyArray b = castToArray(context, a, str(runtime, ...));
```

The problem here is that str is executed every time.  So this form of
castToXXX has a warning in javadocs but it is a potential foot gun.  There
are enough simple string forms where I think this is worth having but it
is undesirable.

I have been considering whether we could pass a lambda for the String part
but that makes me think it would generate a lot more code.

One last observation is that Java instanceof pattern variables encourages
a new idiom of writing code and I am unsure how I feel about it.  Consider
this snippet:

```java
         void put(ThreadContext context, AbstractMemory ptr, long offset, IRubyObject value) {
            if (!(value instanceof Struct)) throw typeError(context, "expected a struct");
            Struct s = (Struct) value;
            byte[] tmp = new byte[Struct.getStructSize(context.runtime, s)];
            s.getMemoryIO().get(0, tmp, 0, tmp.length);
            ptr.getMemoryIO().put(offset, tmp, 0, tmp.length);
         }
```

With pattern var:

```java
         void put(ThreadContext context, AbstractMemory ptr, long offset, IRubyObject value) {
            if (value instanceof Struct s) {
                byte[] tmp = new byte[Struct.getStructSize(context.runtime, s)];
                s.getMemoryIO().get(0, tmp, 0, tmp.length);
                ptr.getMemoryIO().put(offset, tmp, 0, tmp.length);
                return;
            }

            throw typeError(context, "expected a struct");
         }
```

Arguably they are both reasonable ways to write but the pattern var gives
the sense that we are performing a safe cast in a way the compiler will
catch.  The other way is rare to get wrong and it is less weird to see
an error check before the normal path vs the error path appearing to be the
non-special path.  At the end I realized this pattern var will transform
idiomatic Java code to be written using this newer syntactical feature.
We may as well embrace it (sans nice castToXXX methods to hide our casting).

This commit also does a fair amount of style tweaks.  A big part of this new
API work is to see how it "feels".  That lead to some refactoring to help
learn whether the API feels good to me.