Add simple opt when all operands are frozen strings for BuildCompoundString.
If simplifyOperands creates contiguous FrozenString operands it will merge
them.  If it only ends up with a single FrozenString it will change the
instr to a copy.

This still has a bug in that it inplace edits the operands but if interpreter
is still using the original instr it will start printing concat'd contents
until the JIT version takes over.  I will fix in next commit.