Add indy to_s call to specialized dstr call site
When the dynamic elements of a dstring are not trivially
appendable, we must dynamically dispatch to to_s to get a string
result back. However, we don't want to always dispatch to to_s, so
this patch introduces a marker interface Appendable that indicates
there is a fast path for appending certain core types. The indy
call site for dstring will bind each dynamic element to a type
check and branch to the to_s call if necessary, and then proceed
to append each element in turn, either as its Appendable form or
as its to_s result or, eventually, as an anyToString generic form.

This improves performance of dynamic strings containing non-
trivial dynamic elements that must be to_s'ed.

This patch also introduces a second less-specialized path for up
to 5 dynamic elements that avoids using an argument box. This
returns such cases back to the performance of the old non-indy
"pre-inlined" version of this logic and could possibly be expanded
to larger arities as long as the switch does not start to be a
bottleneck.