Handle zero-arg form of dstring
Due to value propagation in IR we may end up with a dstring where
all arguments have propagated as literal strings, leaving this
path with all static strings and no dynamic elements. This will be
turned into a single string load by #8181 but for now this is the
"zero" permutation of the buildString2 logic that has only static
strings.