Allow igv dump to be to stdout
This is not really fantastic as if you dump n scopes for a whole file
then it will dump n XML documents so it cannot just be loaded by
IGV without you slicing out the part you want.  It is still better
than nothing and I didn't want to write something which output
a directory with mangling or autonaming.