Radically rename closures scope name
This will make scope names much easier to glance at to know what they
are.  This also corrects line number from 0-based to 1-based.

CLOSURE_1::CLOSURE_2 is just not super easy to grok.  This new form
will make the closure a combination of the method name using the closure
along with the closure parameters.  For raw lambdas the name will be
->({vars})_{id}.

We still emit closure id onto end of the new name since we can still have
identical closure strings.