Partially undo kwarg optz from #8095
In jruby/jruby#8095 (jruby/jruby@121c60fb) I optimized the JIT to
statically compile in the logic for ruby2_keywords at the time of
JIT. Unfortunately there turned out to be too many cases where
this flag can change at runtime, leading to bugs like jruby/jruby#8119
where a previously-jitted scope gets flipped to ruby2_keywords
later on.

This patch partially reverts that optimization and resumes passing
in the live ruby2_keywords value from the scope. We can revisit
this in the future when it is possible to throw out and re-compile
scopes that get this flag set later on.

Fixes jruby/jruby#8119