Add a property to configure eager constants
In order to avoid polluting the Java:: namespace with classes from
other classloader hierarchies, we are modifying the proxy class
setup process to not eagerly set a constant for newly-encountered
classes. This property controls the behavior, defaulting to true
for JRuby 9.4.x but likely to switch to false in a future release
(probably JRuby 10).

The constants will still be defined if accessed directly, as
before. We just won't define them when we happen to see an object
with an unfamiliar type.

See jruby/jruby#8208 and the bug it fixed jruby/jruby#8156.