Process chdir option sooner so execFillarg sees it
This is a bit hacky due to the way we handle chdir on top of
posix_spawn. The logic for making the command line use sh and cd
lives in execFillarg, since it must handle the case where the
virtual CWD is not the same as the JVM CWD. In order to make sure
that processing also happens for incoming chdir: options, this
patch pre-processes the options hash just for chdir.

Fixes #8126