Implement "chilled" Strings
In implementing the feature I took the least complicated route.  We
get a new Operand and in JIT we just add a second chilled boolean vs
trying to shave this to a three-state value.

One uncertainty is that I mark chilled as both frozen and chilled in
RubyString flags.  I did this because there are lots of isFrozen() checks
in our codebase and I would like to fail fast if we find a path where
chill should convert to mutable and we miss it.  If we were certain we could
probably only set chill flag and then check that in `frozen?`.