Make deprecated methods getlineCall act pre-3.x kwargs behavior
StringIO and possibly other native extensions do not mark their
methods with `keywords=true`.  At some future point they should
do that and call the newer versions of `getlineCall`.

This commit emulates the old behavior in the deprecated methods which
will basically accept that a hash in last position is a kwarg.  This
is not 100% correct but it has not been correct for a long time.  So
this will not make the compatibility any worse but also will not require
a release of stringio in step with JRuby 10.