Optimize literal fixnum and float arrays
This change detects when a literal array contains only fixnums or
floats and uses indy to embed those values in a call site, greatly
reducing the bytecode size for constructing the array. Arrays that
are too large will not fit into the encoded String and prevent
compilation, but they would not likely have fit into bytecode
limits with the old logic anyway.

A benchmark is included.