Rework JVM name mangling for jitted Ruby scopes
This restructures the "mangled" JVM method name used for the JVM
stacktrace and mined for the Ruby backtrace to make it more
readable to normal users.

Instead of "RUBY$method$foo$1"

we get "❤ def foo #1"

The heart is the RUBY marker, def indicates a method, and the # is
the same scope numbering the JIT had before. The delimiter is a
unicode non-breaking space.

This shows in the JVM backtrace similar to these examples:

	at blah.️❤ def foo$1(blah.rb:2)
	at blah.️❤ def bar$2(blah.rb:6)
	at blah.️❤ {} go$6(blah.rb:17)
	at blah.️❤ def times$4(blah.rb:12)
	at blah.️❤ def go$5(blah.rb:17)
	at blah.️❤ {} \=\^main\_$0(blah.rb:20)
	at blah.️❤ def times$4(blah.rb:12)

But still can be parsed by the Ruby backtrace miner to produce the
proper Ruby backtrace elements.

The format of this is negotiable.