Declare variables on first assignment
Naively declaring variables as existing from the beginning of the
method until the end does not seem to be a problem for the JVM,
but debugging tools like IntelliJ's JVM debugger do not appear to
like it. Instead, we add an additional map to track the first
assignment of each local variable and use that as the starting
offset. This leads to some additional labels in the bytecode, but
IntelliJ is now able to properly display all local variables in
its debugger.

Fixes #8256