Set writer modification time to now by default
CRuby does this lazily on first write, when it sets up the
headers, but that laziness is handled inside jzlib for us. Just
set it and if it is updated by user it will be updated in the
writer.

Fixes jruby/jruby#7879, which led to tests being skipped in
rack/rack#2182.