Fix Hash key/value pattern match in JIT
The JIT tries to optimize "eqq" operations ("===" call) to avoid
storing the return value only to immediately load it and check for
truthiness. That works fine in case/when where the eqq is always
followed by exactly one truth test, but for patterns the result of
the eqq is used in other tests later in the code (I think for
delayed determination of error conditions).

This patch adds a flag to EQQInstr to indicate if this was an eqq
used for pattern matching, which has more complicated load/store
requirements than a simple when.

Fixes #8283