3 new Convert API methods
checkToInteger - Same as TypeConverter but it replaces it
integerAsInt - Works from knowing we have RubyInteger and bounds checks to
   make sure we really have an int
integerAsLong - Works from knowing we have RubyInteger and bounds checks to
   make sure we really have a long

Fallout from this is realizing we had two checkToInteger but whatever used
the more generic form which allows to pass in converter is no longer used
for anything but #to_int.  This PR deprecated both versions.

In refactoring to use these I noticed in two places we were not even checking
to see if the value was safely an int/long.  The integerAs* provides this and is
more direct/narrow in what it provides than num2int or num2long.