Properly fall back if can't setAccessible
Changes in e8dbe9c26 broke this fallback by silencing errors from
the old setAccessible (as part of becoming more module-friendly).
This patch checks if the setAccessible was successful and returns
null for any fields that could not be made accessible. This in
turn allows the USABLE fallback to properly reflect actual
usability.