Enhance reflective pid-getting for Windows Java 8+
This is a modified version of a patch proposed in https://github.com/jruby/jruby/issues/7515#issuecomment-1368041882
that uses a third mechanism for getting the PID: the `pid` method
added to ProcessImpl in 8 and made public in Process in 9.

For JRuby 10, which will target Java 17 or 21, we can just call
`pid` directly.

Fixes #7515